apps/basketball-api/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "basketball_api" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/apps/basketball-api/html'
 --cov-report xml:'../../coverage/apps/basketball-api/coverage.xml'
 --html='../../reports/apps/basketball-api/unittests/html/index.html'
 --junitxml='../../reports/apps/basketball-api/unittests/junit.xml'
"""

[tool.poetry]
name = "basketball-api"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "basketball_api"

  [tool.poetry.dependencies]
  python = ">=3.9"
  fastapi = "^0.110.0"
  uvicorn = "^0.27.0"
  sqlalchemy = "^2.0.43"
  alembic = "^1.16.5"
  python-dotenv = "^1.1.1"
  pyhumps = "^3.8.0"
  psycopg2-binary = "^2.9.10"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.3.1"
  flake8 = "7.1.1"
  pytest = "8.3.4"
  pytest-sugar = "1.0.0"
  pytest-cov = "6.0.0"
  pytest-html = "4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
